refactor: extract application runtime into core/runtime.ts
Move the full orchestration logic (subagent init, sandbox, agent
creation loop, hot-reload, heartbeat, cron, Discord transport, API
server, graceful shutdown) from cli.ts main() into a shared
createRuntime() function. cli.ts main() is now ~15 lines: load
config, call createRuntime, wire signal handlers.

Also deduplicates the SIGINT/SIGTERM shutdown handlers that were
copy-pasted as two identical blocks.

cli.ts: 1090 → 708 lines
core/runtime.ts: 371 lines (new)
core/agent-init.ts: 213 lines (from prior commit)
