+++ src/cli.ts
@@ -5,38 +5,16 @@
 import { parseArgs } from "node:util";
 import path from "node:path";
 import { VERSION } from "./version.js";
-import {
-  loadConfig,
-  toAgentConfig,
-  resolveSubagentConfig,
-  resolveSandboxConfigFromFile,
-} from "./core/config.js";
-import { initSubagentBackend, setSubagentSessionStoreFactory } from "./tools/subagent.js";
-import { PiMonoCore } from "./core/pi-mono.js";
-import { DefaultAgentManager } from "./core/agent-manager.js";
-import { DefaultSessionStore } from "./core/session-store.js";
-import { SessionStoreManager } from "./core/session-store-manager.js";
-import { DiscordTransportManager } from "./transports/discord-manager.js";
-import { ThreadBindingManager } from "./core/thread-bindings.js";
+import { loadConfig } from "./core/config.js";
 import { logger } from "./core/logger.js";
-import { LazyTransportContext } from "./tools/reply-react.js";
-import { ProcessRegistry } from "./tools/exec.js";
-import { ContainerManager, SandboxExecutor } from "./sandbox/index.js";
-import { initializeAgent } from "./core/agent-init.js";
+import { createRuntime } from "./core/runtime.js";
 import {
   getConfigPath,
   getIsotopesHome,
   getLogsDir,
-  ensureDirectories,
-  getThreadBindingsPath,
 } from "./core/paths.js";
-import { HotReloadManager } from "./workspace/index.js";
 import { DaemonProcess } from "./daemon/process.js";
 import { ServiceManager, getPlatform, type ServiceConfig } from "./daemon/service.js";
-import { ApiServer } from "./api/server.js";
-import { CronScheduler } from "./automation/cron-job.js";
-import { HeartbeatManager } from "./automation/heartbeat.js";
-import { UsageTracker } from "./core/usage-tracker.js";
 
 // ---------------------------------------------------------------------------
 // Constants
@@ -660,357 +638,22 @@ async function handleLogsCommand(): Promise<void> {
 // ---------------------------------------------------------------------------
 
 async function main() {
-  // Ensure base directories exist
-  await ensureDirectories();
-
-  // Load config from fixed path
   const configPath = values.config ?? getConfigPath();
   logger.info(`Loading config from ${configPath}`);
-
   const config = await loadConfig(configPath);
   logger.info(`Loaded ${config.agents.length} agent(s)`);
 
-  // Single SessionStoreManager backs both main-agent transcripts and
-  // subagent run transcripts. Each store lives at
-  // ~/.isotopes/agents/<normalizedAgentId>/sessions/. See
-  // docs/subagent-architecture.md §4.4.
-  const sessionStoreManager = new SessionStoreManager();
 
-  // Initialize subagent backend with config (M8)
-  if (config.subagent?.enabled) {
-    const subagentConfig = resolveSubagentConfig(config.subagent);
-    initSubagentBackend({
-      permissionMode: subagentConfig.permissionMode,
-      allowedTools: subagentConfig.allowedTools,
-      claude: subagentConfig.claude,
-    });
-    logger.info(`Subagent backend initialized (permissionMode: ${subagentConfig.permissionMode})`);
+  const runtime = await createRuntime({ config, apiPort: getApiPort() });
 
-    // Subagent runs are persisted in their target agent's sessions dir;
-    // the factory resolves the right store on each spawn.
-    setSubagentSessionStoreFactory((agentId) => sessionStoreManager.getOrCreate(agentId));
-    logger.info("Subagent session transcripts → ~/.isotopes/agents/<targetAgentId>/sessions/");
-  }
-
-  // Initialize core with tool registry
-  const core = new PiMonoCore();
-  const agentManager = new DefaultAgentManager(core);
-
-  // Create agents with workspace tools
-  const agentWorkspaces = new Map<string, string>();
-  const transportContexts = new Map<string, LazyTransportContext>();
-  const processRegistries = new Map<string, ProcessRegistry>();
-  const isSingleAgent = config.agents.length === 1;
-
-  // Build a single SandboxExecutor when sandboxing is enabled anywhere.
-  // The agents-level sandbox config (agents.defaults.sandbox or top-level
-  // sandbox) supplies docker/workspaceAccess; per-agent `sandbox` is a
-  // partial override (typically `mode: "off"`). The runtime maintains one
-  // global ContainerManager — per-agent docker overrides are rejected at
-  // config-load time.
-  let sandboxExecutor: SandboxExecutor | undefined;
-  const baseSandboxFile = config.agentDefaults?.sandbox ?? config.sandbox;
-  const resolvedAgentConfigs = config.agents.map((a) =>
-    toAgentConfig(a, config.agentDefaults, config.provider, config.tools, config.compaction, config.sandbox),
-  );
-  const anySandboxed = resolvedAgentConfigs.some((c) => c.sandbox && c.sandbox.mode !== "off");
-  if (anySandboxed) {
-    if (!baseSandboxFile) {
-      throw new Error(
-        "Sandbox is enabled for at least one agent but no agents-level sandbox config was found. " +
-          "Define `agents.defaults.sandbox` or top-level `sandbox` with a docker config.",
-      );
-    }
-    const baseSandbox = resolveSandboxConfigFromFile("<agents-defaults>", undefined, baseSandboxFile);
-    const dockerConfig = baseSandbox?.docker;
-    if (!dockerConfig) {
-      throw new Error("Sandbox is enabled but no docker config could be resolved");
-    }
-    const containerManager = new ContainerManager(dockerConfig);
-    sandboxExecutor = new SandboxExecutor(containerManager, baseSandbox!);
-    logger.info(`Sandbox executor initialized (image: ${dockerConfig.image})`);
-  }
-
-  for (let agentIdx = 0; agentIdx < config.agents.length; agentIdx++) {
-    const agentFile = config.agents[agentIdx];
-
-    // Transport context needs to live in cli.ts for Discord binding
-    const transportCtx = new LazyTransportContext();
-
-    const result = await initializeAgent({
-      agentFile,
-      agentDefaults: config.agentDefaults,
-      provider: config.provider,
-      globalTools: config.tools,
-      compaction: config.compaction,
-      sandbox: config.sandbox,
-      subagent: config.subagent,
-      core,
-      agentManager,
-      isSingleAgent,
-      sandboxExecutor,
-      transportContext: transportCtx,
-    });
-
-    agentWorkspaces.set(result.agentConfig.id, result.workspacePath);
-    transportContexts.set(result.agentConfig.id, transportCtx);
-    processRegistries.set(result.agentConfig.id, result.processRegistry);
-  }
-
-  // Initialize hot-reload for workspace files (M10.5)
-  const hotReload = new HotReloadManager(agentManager, { enabled: true, debounceMs: 500 });
-  for (const [agentId, workspacePath] of agentWorkspaces) {
-    hotReload.register(agentId, workspacePath);
-  }
-  hotReload.onReload((event) => {
-    logger.info(`Hot-reload: workspace reloaded for "${event.agentId}" (${event.changedFiles.join(", ")})`);
-  });
-  hotReload.start();
-  logger.info(`Hot-reload enabled for ${config.agents.length} agent(s)`);
-
-  // Start heartbeat managers for agents with heartbeat config (#191)
-  const heartbeatManagers: HeartbeatManager[] = [];
-
-  for (const agentFile of config.agents) {
-    if (!agentFile.heartbeat?.enabled) continue;
-
-    const workspacePath = agentWorkspaces.get(agentFile.id);
-    if (!workspacePath) continue;
-
-    const hb = new HeartbeatManager({
-      agentId: agentFile.id,
-      workspacePath,
-      config: { ...agentFile.heartbeat, enabled: true },
-      runAgentLoop: async (agentId, prompt, _sessionKey) => {
-        const agent = agentManager.get(agentId);
-        if (!agent) throw new Error(`Agent "${agentId}" not found`);
-
-        let responseText = "";
-        for await (const event of agent.prompt(prompt)) {
-          if (event.type === "text_delta") {
-            responseText += event.text;
-          }
-        }
-        return responseText;
-      },
-    });
-
-    hb.start();
-    heartbeatManagers.push(hb);
-    logger.info(`Heartbeat enabled for "${agentFile.id}" (every ${agentFile.heartbeat.intervalSeconds ?? 300}s)`);
-  }
-
-  // Shared instances for Discord transport and API server
-  const cronScheduler = new CronScheduler();
-  const usageTracker = new UsageTracker();
-
-  // Register cron jobs from config (#193)
-  // Per-agent cron tasks
-  for (const agentFile of config.agents) {
-    if (!agentFile.cron?.tasks?.length) continue;
-
-    for (const task of agentFile.cron.tasks) {
-      cronScheduler.register({
-        name: task.name,
-        expression: task.schedule,
-        agentId: agentFile.id,
-        channelId: task.channel,
-        action: { type: "prompt", prompt: task.prompt },
-        enabled: task.enabled ?? true,
-      });
-    }
-    logger.info(`Registered ${agentFile.cron.tasks.length} cron task(s) for "${agentFile.id}"`);
-  }
-
-  // Top-level cron tasks (from config.cron)
-  if (config.cron?.length) {
-    for (const task of config.cron) {
-      cronScheduler.register({
-        name: task.name,
-        expression: task.expression,
-        agentId: task.agentId,
-        action: task.action,
-        enabled: task.enabled ?? true,
-      });
-    }
-    logger.info(`Registered ${config.cron.length} top-level cron task(s)`);
-  }
-
-  // Wire up cron trigger to prompt agents
-  cronScheduler.onTrigger(async (job) => {
-    const agent = agentManager.get(job.agentId);
-    if (!agent) {
-      logger.error(`Cron job "${job.name}" references unknown agent "${job.agentId}"`);
-      return;
-    }
-
-    let prompt: string;
-    if (job.action.type === "prompt") {
-      prompt = job.action.prompt;
-    } else if (job.action.type === "message") {
-      prompt = job.action.content;
-    } else {
-      logger.warn(`Cron job "${job.name}" has unsupported action type "${job.action.type}" — skipping`);
-      return;
-    }
-
-    const sessionKey = `cron:${job.agentId}:${job.name}`;
-    logger.info(`Cron executing "${job.name}" for agent "${job.agentId}" (session: ${sessionKey})`);
-
-    try {
-      let responseText = "";
-      for await (const event of agent.prompt(prompt)) {
-        if (event.type === "text_delta") {
-          responseText += event.text;
-        }
-      }
-      logger.info(`Cron "${job.name}" completed (${responseText.length} chars)`);
-    } catch (err) {
-      logger.error(`Cron "${job.name}" failed:`, err);
-    }
-  });
-
-  // Start the scheduler
-  cronScheduler.start();
-  if (cronScheduler.listJobs().length > 0) {
-    logger.info(`Cron scheduler started with ${cronScheduler.listJobs().length} job(s)`);
-  }
-
-  // Start Discord transport if configured
-  let discordManager: DiscordTransportManager | undefined;
-  let discordSessionStores: Map<string, DefaultSessionStore> | undefined;
-
-  if (config.channels?.discord) {
-    const accounts = config.channels.discord.accounts ?? {};
-
-    if (Object.keys(accounts).length === 0) {
-      logger.warn("channels.discord present but no accounts configured — skipping");
-    } else {
-      // Pre-warm one store per configured agent via the manager so the
-      // sync sessionStoreForAgent callback always finds a store.
-      const sessionStores = new Map<string, DefaultSessionStore>();
-      for (const agentFile of config.agents) {
-        sessionStores.set(agentFile.id, await sessionStoreManager.getOrCreate(agentFile.id));
-      }
-      discordSessionStores = sessionStores;
-
-      // Pick a default session store: first account's defaultAgentId, else first agent
-      const firstAccount = Object.values(accounts)[0];
-      const defaultAgentId = firstAccount?.defaultAgentId || config.agents[0]?.id || "default";
-      const defaultSessionStore =
-        sessionStores.get(defaultAgentId) ?? (await sessionStoreManager.getOrCreate(defaultAgentId));
-
-      // Create and load persistent thread binding manager (shared across accounts)
-      const threadBindingManager = new ThreadBindingManager({
-        persistPath: getThreadBindingsPath(),
-      });
-      await threadBindingManager.load({ clearStale: true });
-      if (threadBindingManager.size > 0) {
-        logger.info(`Loaded ${threadBindingManager.size} persisted thread binding(s)`);
-      }
-
-      discordManager = new DiscordTransportManager({
-        accounts,
-        shared: {
-          agentManager,
-          sessionStore: defaultSessionStore,
-          sessionStoreForAgent: (agentId) =>
-            sessionStoreManager.peek(agentId) ?? sessionStores.get(agentId) ?? defaultSessionStore,
-          channels: config.channels,
-          threadBindingManager,
-          usageTracker,
-        },
-      });
-
-      const anyThreadBindings = Object.values(accounts).some((a) => a.threadBindings?.enabled);
-      if (anyThreadBindings) {
-        logger.info("Discord thread bindings enabled");
-      }
-
-      await discordManager.start();
-      logger.info(`Discord transport started (${discordManager.size} account(s))`);
-
-      // Bind the first Discord transport to reply/react tools for each agent
-      const firstTransport = discordManager.getAll().values().next().value;
-      if (firstTransport) {
-        for (const [agentId, ctx] of transportContexts) {
-          ctx.setTransport(firstTransport);
-          logger.debug(`Bound Discord transport for reply/react tools (agent: ${agentId})`);
-        }
-      }
-    }
-  }
-
-  // Start API server (REST API)
-  const apiServer = new ApiServer(
-    { port: getApiPort() },
-    cronScheduler,
-    undefined,       // configReloader
-    agentManager,
-    usageTracker,
-    discordSessionStores,
-  );
-  await apiServer.start();
-
-  // Keep process alive
   logger.info("Running... Press Ctrl+C to stop");
 
-  // Graceful shutdown
-  process.on("SIGINT", async () => {
-    logger.info("Shutting down...");
-    cronScheduler.stop();
-    for (const hb of heartbeatManagers) hb.stop();
-    hotReload.stop();
-    if (discordManager) await discordManager.stop();
-    await apiServer.stop();
-
-    // Clean up session stores (#286)
-    sessionStoreManager.destroyAll();
-
-    // Kill orphaned background processes (#286, #289)
-    for (const registry of processRegistries.values()) {
-      registry.clear();
-    }
-
-    // Stop and remove sandbox containers
-    if (sandboxExecutor) {
-      try {
-        await sandboxExecutor.cleanup();
-      } catch (err) {
-        logger.warn(`Sandbox cleanup error: ${err instanceof Error ? err.message : String(err)}`);
-      }
-    }
-
+  const onSignal = async () => {
+    await runtime.shutdown();
     process.exit(0);
-  });
-
-  process.on("SIGTERM", async () => {
-    logger.info("Shutting down...");
-    cronScheduler.stop();
-    for (const hb of heartbeatManagers) hb.stop();
-    hotReload.stop();
-    if (discordManager) await discordManager.stop();
-    await apiServer.stop();
-
-    // Clean up session stores (#286)
-    sessionStoreManager.destroyAll();
-
-    // Kill orphaned background processes (#286, #289)
-    for (const registry of processRegistries.values()) {
-      registry.clear();
-    }
-
-    // Stop and remove sandbox containers
-    if (sandboxExecutor) {
-      try {
-        await sandboxExecutor.cleanup();
-      } catch (err) {
-        logger.warn(`Sandbox cleanup error: ${err instanceof Error ? err.message : String(err)}`);
-      }
-    }
-
-    process.exit(0);
-  });
+  };
+  process.on("SIGINT", onSignal);
+  process.on("SIGTERM", onSignal);
 }
 
 // ---------------------------------------------------------------------------
